requirements.txt
certifi==2021.5.30
chardet==4.0.0
coverage==5.5
idna==2.10
Pillow==8.3.2
pony==0.7.14
psycopg2==2.9.1
requests==2.25.1
six==1.16.0
urllib3==1.26.5
vk-api==11.9.4

